# openapi_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "\u623f\u5c4b\u79df\u8d41AI\u8425\u9500\u7cfb\u7edf",
        "description": "AI\u9a71\u52a8\u7684\u623f\u6e90\u6587\u6848\u751f\u6210\u4e0e\u591a\u5e73\u53f0\u53d1\u5e03\u7cfb\u7edf",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/scripts/generate": {
            "post": {
                "tags": [
                    "\u6587\u6848\u751f\u6210"
                ],
                "summary": "Generate Script",
                "description": "\u751f\u6210AI\u6587\u6848\n\n- **house_id**: \u623f\u6e90ID\n- **template_style**: \u6a21\u677f\u98ce\u683c\uff08professional/friendly/urgent\uff09\n\n\u8fd4\u56de\uff1a\u751f\u6210\u7684\u6587\u6848\u5bf9\u8c61",
                "operationId": "generate_script_api_v1_scripts_generate_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScriptGenerateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScriptResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/scripts/{script_id}": {
            "get": {
                "tags": [
                    "\u6587\u6848\u751f\u6210"
                ],
                "summary": "Get Script",
                "description": "\u83b7\u53d6\u6587\u6848\u8be6\u60c5\n\n- **script_id**: \u6587\u6848ID\n\n\u8fd4\u56de\uff1a\u6587\u6848\u8be6\u7ec6\u4fe1\u606f",
                "operationId": "get_script_api_v1_scripts__script_id__get",
                "parameters": [
                    {
                        "name": "script_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Script Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScriptResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "\u6587\u6848\u751f\u6210"
                ],
                "summary": "Update Script",
                "description": "\u66f4\u65b0\u6587\u6848\uff08\u7f16\u8f91\u540e\u4fdd\u5b58\uff09\n\n- **script_id**: \u6587\u6848ID\n- **request**: \u66f4\u65b0\u5185\u5bb9\uff08title, body, tags\uff09\n\n\u8fd4\u56de\uff1a\u66f4\u65b0\u540e\u7684\u6587\u6848\u5bf9\u8c61",
                "operationId": "update_script_api_v1_scripts__script_id__put",
                "parameters": [
                    {
                        "name": "script_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Script Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ScriptUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScriptResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/scripts": {
            "get": {
                "tags": [
                    "\u6587\u6848\u751f\u6210"
                ],
                "summary": "Get Scripts",
                "description": "\u83b7\u53d6\u6587\u6848\u5217\u8868\n\n- **house_id**: \u7b5b\u9009\u6307\u5b9a\u623f\u6e90\u7684\u6587\u6848\uff08\u53ef\u9009\uff09\n- **skip**: \u8df3\u8fc7\u8bb0\u5f55\u6570\n- **limit**: \u8fd4\u56de\u8bb0\u5f55\u6570\n\n\u8fd4\u56de\uff1a\u6587\u6848\u5217\u8868\u548c\u603b\u6570",
                "operationId": "get_scripts_api_v1_scripts_get",
                "parameters": [
                    {
                        "name": "house_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "House Id"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ScriptListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ScriptGenerateRequest": {
                "properties": {
                    "house_id": {
                        "type": "integer",
                        "title": "House Id",
                        "description": "\u623f\u6e90ID"
                    },
                    "template_style": {
                        "type": "string",
                        "title": "Template Style",
                        "description": "\u6a21\u677f\u98ce\u683c\uff08professional/friendly/urgent\uff09",
                        "default": "professional"
                    }
                },
                "type": "object",
                "required": [
                    "house_id"
                ],
                "title": "ScriptGenerateRequest",
                "description": "\u6587\u6848\u751f\u6210\u8bf7\u6c42Schema\n\u7528\u4e8e\uff1a\u8bf7\u6c42AI\u751f\u6210\u6587\u6848"
            },
            "ScriptListResponse": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/ScriptResponse"
                        },
                        "type": "array",
                        "title": "Items"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "total"
                ],
                "title": "ScriptListResponse",
                "description": "\u6587\u6848\u5217\u8868\u54cd\u5e94Schema"
            },
            "ScriptResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "house_id": {
                        "type": "integer",
                        "title": "House Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "body": {
                        "type": "string",
                        "title": "Body"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "default": []
                    },
                    "platform": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Platform"
                    },
                    "template_style": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Template Style"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "house_id",
                    "title",
                    "body",
                    "created_at"
                ],
                "title": "ScriptResponse",
                "description": "\u6587\u6848\u54cd\u5e94Schema\n\u7528\u4e8e\uff1a\u8fd4\u56de\u6587\u6848\u8be6\u60c5"
            },
            "ScriptUpdateRequest": {
                "properties": {
                    "title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Title",
                        "description": "\u6587\u6848\u6807\u9898"
                    },
                    "body": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Body",
                        "description": "\u6587\u6848\u6b63\u6587"
                    },
                    "tags": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tags",
                        "description": "\u6807\u7b7e\u5217\u8868"
                    }
                },
                "type": "object",
                "title": "ScriptUpdateRequest",
                "description": "\u6587\u6848\u66f4\u65b0\u8bf7\u6c42Schema\n\u7528\u4e8e\uff1a\u7f16\u8f91\u6587\u6848\u540e\u4fdd\u5b58"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}